+++ tests/Database/Postgres/DatabaseEloquentBuilderTest.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Illuminate\Tests\Database\Postgres;
+
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Query\Builder as BaseBuilder;
+use Illuminate\Support\Carbon;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+
+class DatabaseEloquentBuilderTest extends TestCase
+{
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+
+        Carbon::setTestNow(null);
+
+        m::close();
+    }
+
+    public function testOperatorsWhere()
+    {
+        $builder = $this->getBuilder();
+        $builder->getQuery()->shouldReceive('where')->once()->with('foo', '@>', 'bar');
+        $result = $builder->where('foo', '@>', 'bar');
+        $this->assertEquals($result, $builder);
+    }
+
+    protected function getBuilder()
+    {
+        return new Builder($this->getMockQueryBuilder());
+    }
+
+    protected function getMockQueryBuilder()
+    {
+        $query = m::mock(BaseBuilder::class);
+        $query->shouldReceive('from')->with('foo_table');
+
+        return $query;
+    }
+}
